worker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "citylens-engine-worker"
version = "0.1.0"
description = "Citylens Engine Worker (Cloud Run Job)"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Citylens" }]
dependencies = [
  "numpy>=1.23",
  "pillow>=10.0",
  "pydantic>=2.0",
  "google-cloud-firestore>=2.11",
  "google-cloud-storage>=2.14",
  "google-auth>=2.25",
  "requests>=2.31",
  "fiona>=1.9",
  "pyproj>=3.0",
  "rasterio>=1.3",
  "shapely>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.3"]

# citylens-core dependency strategy:
# - Production: install from git in Dockerfile (see worker/Dockerfile)
# - Local dev: `pip install -e ../citylens-core`

[tool.setuptools]
py-modules = ["worker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
